+++ src/code/dialogs/checklist.js
@@ -18,6 +18,7 @@ import { getSelectedOperation } from "../selectedOp";
 import wX from "../wX";
 
 import PortalUI from "../ui/portal";
+import LinkUI from "../ui/link";
 
 const OperationChecklistDialog = WDialog.extend({
   statics: {
@@ -129,7 +130,7 @@ const OperationChecklistDialog = WDialog.extend({
         sort: (a, b) => a.localeCompare(b),
         format: (cell, value, thing) => {
           if (thing instanceof WasabeeLink) {
-            cell.appendChild(thing.displayFormat(operation));
+            cell.appendChild(LinkUI.displayFormat(thing, operation));
             if (this._smallScreen) cell.colSpan = 2;
           } else {
             cell.appendChild(
@@ -362,7 +363,7 @@ const OperationChecklistDialog = WDialog.extend({
         for (const [link, c] of emptyFieldLinks) {
           const li = L.DomUtil.create("li", "empty-field-link", content);
           li.textContent = c;
-          li.appendChild(link.displayFormat(operation));
+          li.appendChild(LinkUI.displayFormat(link, operation));
         }
         alert(container, true);
       } else {
