+++ src/code/dialogs/mergeDialog.js
@@ -7,6 +7,7 @@ import { getSelectedOperation, makeSelectedOperation } from "../selectedOp";
 import { drawBackgroundOp } from "../mapDrawing";
 
 import PortalUI from "../ui/portal";
+import LinkUI from "../ui/link";
 
 const MergeDialog = WDialog.extend({
   statics: {
@@ -231,7 +232,7 @@ const MergeDialog = WDialog.extend({
         format: async (cell, value, e) => {
           const op = e.type === "-" ? origin : operation;
           if (e.data.type === "link") {
-            cell.appendChild(e.data.link.displayFormat(op));
+            cell.appendChild(LinkUI.displayFormat(e.data.link, op));
           } else if (e.data.type === "portal") {
             cell.appendChild(PortalUI.displayFormat(e.data.portal));
           } else if (e.data.type === "marker") {
